restructure how the dash app is created (#117)
* restructure how the dash app is created

* update docs on the various ways in which dashboard can be run

* modularize imports

* allow page size to be specified

* allow assets to be accessible from views

* pr nits
