+++ rubicon_ml/ui/dashboard.py
@@ -3,13 +3,19 @@ import threading
 
 import dash_bootstrap_components as dbc
 import dash_html_components as html
+from dash import Dash
 
-from rubicon_ml.ui.app import app
+from rubicon_ml.ui.callbacks import (
+    set_project_explorer_callbacks,
+    set_project_selection_callbacks,
+)
 from rubicon_ml.ui.model import RubiconModel
-from rubicon_ml.ui.views.footer import make_footer_layout
-from rubicon_ml.ui.views.header import make_header_layout
-from rubicon_ml.ui.views.project_explorer import make_project_explorer_layout
-from rubicon_ml.ui.views.project_selection import make_project_selection_layout
+from rubicon_ml.ui.views import (
+    make_footer_layout,
+    make_header_layout,
+    make_project_explorer_layout,
+    make_project_selection_layout,
+)
 
 
 class Dashboard:
@@ -26,17 +32,29 @@ class Dashboard:
     page_size : int, optional
         The number of rows that will be displayed on a page within the
         experiment table.
+    dash_options: dict, optional
+        Additional arguments specific to the Dash app. Visit the
+        `docs <https://dash.plotly.com/reference>`_ to see what's
+        available. Note, `requests_pathname_prefix` is useful for proxy
+        troubles.
     storage_options : dict, optional
         Additional keyword arguments specific to the protocol being chosen. They
         are passed directly to the underlying filesystem class.
     """
 
-    def __init__(self, persistence, root_dir=None, page_size=10, **storage_options):
+    def __init__(
+        self,
+        persistence,
+        root_dir=None,
+        page_size=10,
+        dash_options={},
+        **storage_options,
+    ):
         self.rubicon_model = RubiconModel(persistence, root_dir, **storage_options)
 
-        self._app = app
-        self._app._page_size = page_size
+        self._app = Dash(__name__, title="Rubicon", **dash_options)
         self._app._rubicon_model = self.rubicon_model
+        self._app._page_size = page_size
         self._app.layout = html.Div(
             [
                 dbc.Row(make_header_layout()),
@@ -57,6 +75,9 @@ class Dashboard:
                 dbc.Row(make_footer_layout()),
             ]
         )
+
+        set_project_selection_callbacks(self._app)
+        set_project_explorer_callbacks(self._app)
 
     def run_server(self, **kwargs):
         """Serve the dash app on an external web page.
@@ -103,6 +124,6 @@ class Dashboard:
         running_server_thread.daemon = True
         running_server_thread.start()
 
-        proxied_host = os.path.join(host, app.config["requests_pathname_prefix"].lstrip("/"))
+        proxied_host = os.path.join(host, self._app.config["requests_pathname_prefix"].lstrip("/"))
 
         return IFrame(proxied_host, **i_frame_kwargs)
