+++ rubicon_ml/ui/views/__init__.py
@@ -1,0 +1,11 @@
+from .footer import make_footer_layout
+from .header import make_header_layout
+from .project_explorer import make_project_explorer_layout
+from .project_selection import make_project_selection_layout
+
+__all__ = [
+    "make_footer_layout",
+    "make_header_layout",
+    "make_project_explorer_layout",
+    "make_project_selection_layout",
+]
