+++ rubicon_ml/ui/views/project_selection.py
@@ -1,8 +1,4 @@
-import dash_bootstrap_components as dbc
 import dash_html_components as html
-from dash.dependencies import Input, Output
-
-from rubicon_ml.ui.app import app
 
 
 def make_project_selection_layout():
@@ -25,26 +21,4 @@ def make_project_selection_layout():
             ),
             html.Div(id="project-selection--list", className="project-selection--list"),
         ],
-    )
-
-
-@app.callback(
-    Output("project-selection--list", "children"),
-    [Input("project-selection--refresh-projects-btn", "n_clicks")],
-)
-def _update_project_options(n_clicks):
-    """Use the Rubicon client to load the available projects."""
-    app._rubicon_model.update_projects()
-
-    return dbc.ListGroup(
-        children=[
-            dbc.ListGroupItem(
-                id={"type": "project-selection--project", "index": project["value"]},
-                className="project-selection--project",
-                children=project["value"],
-                n_clicks=0,
-                action=True,
-            )
-            for project in app._rubicon_model.project_options
-        ]
     )
